backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
jira==3.6.0
python-dotenv==1.0.0
pandas==2.2.0
sqlalchemy==2.0.25
apscheduler==3.10.4
pydantic==2.5.3
